fdb: cast scoped FdbDest enum for %d format & restore p4orch mock
- Add static_cast<int>(...) at remaining FdbDest %d log sites in
  fdborch.cpp and qualify FdbDest:: in mock_tests/fdborch/* and
  mock_tests/fdbsyncd/* (required after enum class conversion).
- Restore orchagent/p4orch/tests/mock_routeorch.h (lost during rebase)
  and update its RouteOrch ctor declaration to match the current
  prototype (adds NeighOrch* and ZmqServer*).

+++ orchagent/fdborch.cpp
@@ -720,7 +720,7 @@ void FdbOrch::update(sai_fdb_event_t        type,
         /* Remove local neighbor entry if exists
          */
         SWSS_LOG_INFO("Received mac age out for mac:%s vlan:0x%" PRIx64 "of type:%d",
-				update.entry.mac.to_string().c_str(), update.entry.bv_id, existing_entry->second.dest_type);
+				update.entry.mac.to_string().c_str(), update.entry.bv_id, static_cast<int>(existing_entry->second.dest_type));
 
         gNeighOrch->processFDBResolve(update.entry);
